Crop large images into patches
Due to the limitation of GPU, sometimes it cannot run a single image with high resolution. Therefore, we need to crop the image into ptaches, processing them one by one, and reconstruct the processed image from these patches.
The crop size and crop stride can be set freely. The algorithm firstly calculate the coordinates of crop points, and then crop the image to patches.